+++ src/AutoRest.CSharp/Common/Output/Models/Types/SystemObjectType.cs
@@ -11,7 +11,6 @@ using System.Reflection;
 using AutoRest.CSharp.Common.Utilities;
 using AutoRest.CSharp.Generation.Types;
 using AutoRest.CSharp.Input;
-using AutoRest.CSharp.Mgmt.Decorator;
 using AutoRest.CSharp.Output.Models.Requests;
 using AutoRest.CSharp.Output.Models.Shared;
 using Azure.Core;
