+++ src/AutoRest.CSharp/LowLevel/Decorator/TypeReferenceTypeChooser.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using AutoRest.CSharp.AutoRest.Plugins;
+using AutoRest.CSharp.Common.Input;
 using AutoRest.CSharp.Common.Utilities;
 using AutoRest.CSharp.Generation.Types;
 using AutoRest.CSharp.Input;
@@ -18,7 +19,7 @@ using AutoRest.CSharp.Output.Models.Types;
 using Azure.ResourceManager;
 using Azure.ResourceManager.Resources.Models;
 
-namespace AutoRest.CSharp.Mgmt.Decorator
+namespace AutoRest.CSharp.Decorator
 {
     /// <summary>
     /// This is a utility class to check and keep the result of whether a <c>MgmtObjectType</c> class can be replaced by
@@ -29,16 +30,16 @@ namespace AutoRest.CSharp.Mgmt.Decorator
         private static IReadOnlyList<System.Type>? _typeReferenceTypes;
         private static IReadOnlyList<System.Type> TypeReferenceTypes => _typeReferenceTypes ??= ReferenceClassFinder.GetTypeReferenceTypes();
 
-        private static ConcurrentDictionary<Schema, CSharpType?> _valueCache = new ConcurrentDictionary<Schema, CSharpType?>();
+        private static ConcurrentDictionary<InputModelType, CSharpType?> _valueCache = new ConcurrentDictionary<InputModelType, CSharpType?>();
 
         /// <summary>
         /// Check whether a <c>MgmtObjectType</c> class can be replaced by an external type, and return the external type if available.
         /// </summary>
         /// <param name="typeToReplace">Type to check</param>
         /// <returns>Matched external type or null if not found</returns>
-        public static CSharpType? GetExactMatch(MgmtObjectType typeToReplace)
+        public static CSharpType? GetExactMatch(ModelTypeProvider typeToReplace)
         {
-            if (_valueCache.TryGetValue(typeToReplace.ObjectSchema, out var result))
+            if (_valueCache.TryGetValue(typeToReplace.InputModel, out var result))
                 return result;
 
             foreach (System.Type replacementType in TypeReferenceTypes)
@@ -46,23 +47,13 @@ namespace AutoRest.CSharp.Mgmt.Decorator
                 if (PropertyMatchDetection.IsEqual(replacementType, typeToReplace))
                 {
                     var csharpType = CSharpType.FromSystemType(MgmtContext.Context, replacementType);
-                    _valueCache.TryAdd(typeToReplace.ObjectSchema, csharpType);
+                    _valueCache.TryAdd(typeToReplace.InputModel, csharpType);
                     return csharpType;
                 }
             }
 
-            _valueCache.TryAdd(typeToReplace.ObjectSchema, null);
+            _valueCache.TryAdd(typeToReplace.InputModel, null);
             return null;
-        }
-
-        /// <summary>
-        /// Check whether there is a match for the given schema.
-        /// </summary>
-        /// <param name="schema"><c>ObjectSchema</c> of the target type</param>
-        /// <returns></returns>
-        public static bool HasMatch(ObjectSchema schema)
-        {
-            return _valueCache.TryGetValue(schema, out var match) && match != null;
         }
     }
 }
